syscalls/pty08: stress sysfs vc nodes
Attempt to stress as much as possible the ssfs v nodes. This has been
the only semi-reliable way to reproduced a list corruption using a
PREEMPT_RT kernel on aarch64.

Note: The pty are only a convenient way to trigger pressure on the
sysfs creating/removing lots of nodes fast. The crash actually happens
in systemd-udevd when issuing syscalls that trigger a dentry pathwalk.

Signed-off-by: Eric Chanudet <echanude@redhat.com>