fs/dentry: attempt to stress dentry lookups
Use openat/renameat/unlinkat in parallel to try and stress the
operations on a dentry d_subdir list and trigger a corruption. This is
an attempt at reproducing more consistently a bug that was initially
triggered by pty07 when multiple systemd-udevd child hit the virtual
console sysfs and /run/udev/{data,watch} directory until a list
corruption was observed.

Signed-off-by: Eric Chanudet <echanude@redhat.com>